Move validate-target to utils
