+++ src/shared/utils/validate-target.ts
